Improve StackName/Version discovery for ASG and ELB
Figuring out the stack name/version of an AWS object from its name
seems to be a tricky business: the format of the name uses dashes
and we use them to separate name and version, moreover a user may
use a dash as a part of the stack version.

Instead of trying to split the combined string on dash, we should
use Tags.