Don't describe ASG instances when there are none
When a new AutoScaling Group is created (or when a stack is deleted),
for a short period of time it doesn't have any instances.  Then the
agent tries to call describe_instances w/o the list of instance ids,
which returns the list of *all* account instances.

Fix this by not describing instances, then the ASG's list is empty.