Add PostgreSQL database entities from (static) PostgreSQL clusters
Listing database names requires the user and password to be provided.  These
can be given on the command line or they will be taken from the environment:
AGENT_POSTGRESQL_USER and AGENT_POSTGRESQL_PASS.

If database credentials are not specified, the agent will not add or remove
entities of type 'postgresql_database'.